# docs/swagger.yaml
basePath: /
definitions:
...
  dto.ItemCreateRequest:
    properties:
      code:
        example: crear_usuario
        maxLength: 50
        minLength: 3
        type: string
      name:
        example: Crear Usuario
        maxLength: 50
        minLength: 3
        type: string
    required:
    - code
    - name
    type: object
  dto.ItemResponse:
    properties:
      code:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  dto.ItemUpdateRequest:
    properties:
      code:
        example: actualizar_usuario
        maxLength: 50
        minLength: 3
        type: string
      name:
        example: Actualizar Usuario
        maxLength: 50
        minLength: 3
        type: string
    required:
    - code
    - name
    type: object
  dto.LoginRequest:
    properties:
      email:
        example: cesar@tudominio.com
        type: string
      password:
        description: ← Placeholder genérico
        example: '********'
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
  dto.LoginResponse:
    properties:
      user:
        $ref: '#/definitions/dto.UserResponse'
    type: object
  dto.LogoutResponse:
    properties:
      message:
        type: string
    type: object
  dto.MeResponse:
    properties:
      user:
        $ref: '#/definitions/dto.UserResponse'
    type: object
...
  dto.ProfileSummary:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
...
  dto.RefreshResponse:
    properties:
      message:
        type: string
    type: object
...
  dto.UserModuleResponse:
    properties:
      items:
        description: ← Items con code incluido
        items:
          $ref: '#/definitions/dto.ItemResponse'
        type: array
      name:
        type: string
      slug:
        type: string
    type: object
  dto.UserResponse:
    properties:
      date:
        type: string
      email:
        type: string
      id:
        type: integer
      modules:
        description: ← Usa el tipo específico
        items:
          $ref: '#/definitions/dto.UserModuleResponse'
        type: array
      name:
        type: string
      phone:
        type: string
      profile:
        $ref: '#/definitions/dto.ProfileSummary'
      profile_id:
        type: integer
      state:
        type: integer
      time:
        type: string
    type: object
...
host: localhost:8082
info:
  contact: {}
  description: API Backend para GOLANG DEMO.
  title: Golang DEMO
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Autentica al usuario y crea una sesión con cookie HTTP-only
      parameters:
      - description: Credenciales
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.LoginResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Iniciar sesión
      tags:
      - Auth
  /auth/logout:
    post:
      description: Invalida la sesión actual y elimina la cookie
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.LogoutResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Cerrar sesión
      tags:
      - Auth
  /auth/me:
    get:
      description: Retorna los datos del usuario actualmente autenticado, incluyendo
        su perfil, estado y módulos con permisos (items) asignados
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MeResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Obtener usuario autenticado
      tags:
      - Auth
  /auth/refresh:
    post:
      description: Extiende el TTL de la sesión actual
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.RefreshResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Renovar sesión
      tags:
      - Auth
...
  /items:
    get:
      description: Retorna todos los items registrados ordenados por ID descendente
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.ItemResponse'
            type: array
      summary: Listar items
      tags:
      - Items
    post:
      consumes:
      - application/json
      description: Crea un nuevo item (acción) con nombre y código únicos
      parameters:
      - description: Datos del item
        in: body
        name: item
        required: true
        schema:
          $ref: '#/definitions/dto.ItemCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.ItemResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Crear nuevo item
      tags:
      - Items
  /items/{id}:
    delete:
      description: Elimina un item por su ID (valida dependencias con profile_module_item)
      parameters:
      - description: ID del item
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Eliminar item
      tags:
      - Items
    get:
      description: Retorna un item específico por su ID
      parameters:
      - description: ID del item
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ItemResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Obtener item por ID
      tags:
      - Items
    put:
      consumes:
      - application/json
      description: Actualiza el nombre y código de un item existente
      parameters:
      - description: ID del item
        in: path
        name: id
        required: true
        type: integer
      - description: Nuevos datos
        in: body
        name: item
        required: true
        schema:
          $ref: '#/definitions/dto.ItemUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ItemResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Actualizar item
      tags:
      - Items
...
schemes:
- http
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"